requirements.txt
Flask
docker
GitPython